+++ driver/lib/sessions/log-monitor.ts
@@ -1,10 +1,14 @@
 import type {EventEmitter} from 'node:events';
 import { log as logger } from '../logger';
+import { retryInterval } from 'asyncbox';
 export interface LogEntry {
   timestamp: number;
   level: string,
   message: string;
 }
+
+const DEFAULT_MAX_RETRY_COUNT = 20;
+const DEFAULT_BACKOFF_TIME_MS = 3000;
 
 export type Filter = (x: LogEntry) => Promise<boolean>;
 
@@ -18,7 +22,6 @@ export class LogMonitor {
     this._logsEmitter = logsEmitter;
     this._outputListener = null;
     this._filter = filter;
-    this._lastMatch = null;
   }
 
   get started(): boolean {
@@ -32,6 +35,26 @@ export class LogMonitor {
   get lastMatch(): LogEntry | null {
     return this._lastMatch;
   }
+
+  async waitForLastMatchExist(
+    maxRetryCount: number = DEFAULT_MAX_RETRY_COUNT,
+    retryBackoffTime: number = DEFAULT_BACKOFF_TIME_MS,
+  ): Promise<LogEntry | null> {
+    return await retryInterval(
+      maxRetryCount,
+      retryBackoffTime,
+      async () => {
+        if (this._lastMatch !== null) {
+          return this._lastMatch;
+        }
+        throw new Error(
+          `No matched log found with ${retryBackoffTime} ms interval ` +
+          `up to ${maxRetryCount} times. Increasing appium:retryBackoffTime ` +
+          `and appium:maxRetryCount would help.`
+        );
+      },
+    );
+  };
 
   start(): this {
     if (this.started) {