+++ driver/lib/sessions/session.ts
@@ -20,7 +20,7 @@ export const reConnectFlutterDriver = async function(this: FlutterDriver, caps: 
       this.socket = await connectIOSSession.bind(this)(this.proxydriver, caps, true);
       break;
     case PLATFORM.ANDROID:
-      this.socket = await connectAndroidSession.bind(this)(this.proxydriver, caps);
+      this.socket = await connectAndroidSession.bind(this)(this.proxydriver, caps, true);
       break;
     default:
       this.log.errorAndThrow(